+++ lib/vagrant_utm/driver/version_4_5.rb
@@ -174,9 +174,8 @@ module VagrantPlugins
         end
 
         def read_guest_ip
-          command = ["read_guest_ip.applescript", @uuid]
-          output = execute_osa_script(command)
-          output.strip
+          output = execute("ip-address", @uuid)
+          output.strip.split("\n")
         end
 
         def read_network_interfaces
@@ -193,6 +192,14 @@ module VagrantPlugins
           end
 
           nics
+        end
+
+        def set_mac_address(mac) # rubocop:disable Naming/AccessorMethodName
+          # Set the MAC address of the first NIC (index 0)
+          # Set MAC address to given value or randomize it if nil
+          mac = random_mac_address if mac.nil?
+          command = ["set_mac_address.applescript", @uuid, "0", mac]
+          execute_osa_script(command)
         end
 
         def ssh_port(expected_port)