+++ spec/support/shared/utm_context.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-shared_context "utm" do
+shared_context "utm" do # rubocop:disable Metrics/BlockLength
   include_context "unit"
 
   let(:utm_context) { true }
@@ -8,6 +8,7 @@ shared_context "utm" do
   let(:utm_version) { "4.5.0" }
   let(:subprocess) { double("Vagrant::Util::Subprocess") }
   let(:script_path) { File.expand_path("../../../lib/vagrant_utm/scripts/", __dir__) }
+  let(:utmctl_path) { "/usr/local/bin/utmctl" }
 
   # this is a helper that returns a duck type suitable from a system command
   # execution; allows setting exit_code, stdout, and stderr in stubs.